When someone updates their own profile the Student and Instructor tables update appropriately.
If the user has not been made into a Student yet, they will be (as every user should be at least a Student under this design).  If they have been a Student or an Instructor the appropriate tables will be updated with their profile information.

Note that there is a limitation in that an Administrator currently cannot edit a user's profile.

Signed-off-by: ecnalyr <ecnalyr@gmail.com>