Added the foundation framework for the notification system; misc bug fixes.
Added the notification model.  When an Instructor applies to teach a Course a notification is generated for an Admin, with a link, to attend to the newly-applied-for Course.  Fixed bugs involving rare situations where a user may already be in the "role" of Student or Instructor but has not been added to the Student or Instructor 'database table' .

Signed-off-by: ecnalyr <ecnalyr@gmail.com>