Paypal works
Very difficult to test as the site basically has to be live - but it
definitely works.  When a student views /Attendance/StudentDetail/8 they
get a message stating whether or not they have paid for the course.  If
they have not paid for the course they are given a Paypal button that
allows them to make the payment.  Once they make the payment an IPN call
is made and the user's payment information is verified.  Upon
verification the associated Enrollment.Paid is changed from False to
True.