Removed Duration from Course model, changed it to use StartDate and EndDate
StartTime and EndTime are determined by the full value of StartDate and
EndDate.  During Course cration the user must select a full StartTime
and a full EndTime for the values to be correct - this is explained in
the view.

Also changed default Attendance value to true (or present), and adjusted
some View formatting.